Enable locator feature from vite plugin
