+++ configs/tsup.config.ts
@@ -29,7 +29,6 @@ export default ({
     const options: Options = {
       watch: config.watch,
       clean: config.watch ? false : true,
-      minify: config.watch ? false : true,
       treeshake: config.watch ? false : true,
       dts: {
         entry: [entry, ...mappedAdditionalEntries],
