+++ examples/sandbox/src/devtools.ts
@@ -1,9 +1,4 @@
 import { attachDevtoolsOverlay } from '@solid-devtools/overlay'
-import { useLocator } from '@solid-devtools/debugger'
-
-useLocator({
-  targetIDE: 'vscode',
-})
 
 !process.env.EXT &&
   attachDevtoolsOverlay({
