+++ packages/debugger/package.json
@@ -70,7 +70,6 @@
   },
   "dependencies": {
     "@solid-devtools/shared": "workspace:^0.10.3",
-    "@solid-devtools/transform": "workspace:^0.9.0",
     "@solid-primitives/bounds": "^0.0.105",
     "@solid-primitives/cursor": "^0.0.103",
     "@solid-primitives/event-bus": "^0.1.3",
