+++ packages/debugger/src/locator/markComponent.ts
@@ -1,6 +1,6 @@
 import { getOwner, getOwnerType } from '../main/utils'
-import { LocationAttr, MARK_COMPONENT_FN_NAME } from '@solid-devtools/transform/types'
 import { NodeType, Solid } from '../types'
+import { LocationAttr } from './findComponent'
 
 export function markComponentLoc(location: LocationAttr): void {
   let owner = getOwner()
@@ -9,6 +9,3 @@ export function markComponentLoc(location: LocationAttr): void {
   if (type === NodeType.Component) (owner as Solid.Component).location = location
   else if (type === NodeType.Refresh) (owner.owner as Solid.Component).location = location
 }
-
-// transform uses this global function to mark components
-;(globalThis as any)[MARK_COMPONENT_FN_NAME] = markComponentLoc
