+++ packages/debugger/src/types.ts
@@ -1,14 +1,6 @@
-export type { LocationAttr } from '@solid-devtools/transform/types'
-export type { EncodedValue, EncodedValueMap } from './inspector/serialize'
-export { ValueType, INFINITY, NAN, NEGATIVE_INFINITY, UNDEFINED } from './inspector/serialize'
 export * from './main/types'
 export * from './main/constants'
-export type {
-  InspectorUpdate,
-  SetInspectedNodeData,
-  ToggleInspectedValueData,
-  ProxyPropsUpdate,
-  StoreNodeUpdate,
-  ValueNodeUpdate,
-  HighlightElementPayload,
-} from '.'
+
+export * from './inspector/types'
+
+export * from './locator/types'
