+++ packages/transform/package.json
@@ -33,23 +33,11 @@
   "module": "dist/index.js",
   "types": "dist/index.d.ts",
   "exports": {
-    ".": {
-      "import": "./dist/index.js",
-      "require": "./dist/index.cjs",
-      "types": "./dist/index.d.ts"
+    "import": {
+      "types": "./dist/index.d.ts",
+      "default": "./dist/index.js"
     },
-    "./types": {
-      "import": "./dist/types.js",
-      "require": "./dist/types.cjs",
-      "types": "./dist/types.d.ts"
-    }
-  },
-  "typesVersions": {
-    "*": {
-      "types": [
-        "./dist/types.d.ts"
-      ]
-    }
+    "require": "./dist/index.cjs"
   },
   "scripts": {
     "dev": "tsup --watch",
@@ -75,10 +63,11 @@
     "@babel/core": "^7.19.6",
     "@babel/plugin-syntax-typescript": "^7.18.6",
     "@babel/types": "^7.19.4",
-    "@solid-devtools/shared": "workspace:^0.10.1"
+    "@solid-devtools/shared": "workspace:^0.10.1",
+    "@solid-devtools/debugger": "workspace:^0.17.0"
   },
   "peerDependencies": {
     "solid-js": "^1.6.2",
-    "vite": "^2.2.3 || ^3.0.0"
+    "vite": "^2.2.3 || ^3.0.0 || ^4.0.0"
   }
 }
