--- packages/transform/src/types.ts
@@ -1,6 +1,0 @@
-export const MARK_COMPONENT_FN_NAME = `_$markComponentLoc`
-
-export const WINDOW_PROJECTPATH_PROPERTY = '$sdt_projectPath'
-export const LOCATION_ATTRIBUTE_NAME = 'data-source-loc'
-
-export type LocationAttr = `${string}:${number}:${number}`
