+++ packages/transform/tsup.config.ts
@@ -1,5 +1,3 @@
 import defineConfig from '../../configs/tsup.config'
 
-export default defineConfig({
-  additionalEntries: ['types'],
-})
+export default defineConfig({})
